chore: moved common utilities into @morfeo/utils
