+++ packages/core/src/index.ts
@@ -1,7 +1,6 @@
 export * from './theme';
 export * from './types';
-export * from './parsers';
-export * from './utils/deepMerge';
 export * from './morfeo';
+export * from './parsers';
 
 export * from '@morfeo/spec';
