+++ packages/utils/src/index.ts
@@ -1,1 +1,3 @@
 export * from './strings';
+export * from './deepMerge';
+export * from './parseNumber';
